+++ tools/liveblocks-cli/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## vNEXT (not yet released)
 
-## v1.2.1
+## v1.3.0
 
+- Add feeds support (`feeds:write` permission)
+- Add verbose logging toggle
+- Fix permission validation to accept all valid permission combinations
 - Support passing extra arguments to `--cmd` (`-c`), appended to the command or
   replacing `{}` if present